+++ Lab6_trigger_procedure/src/main/java/com/api/repository/RatingCustomRepository.java
@@ -1,5 +1,5 @@
 package com.api.repository;
 
 public interface RatingCustomRepository {
-    Float getAvgRating();
+    double getAvgRating();
 }
